Move apollo-client to /server
