+++ bin/fragments.js
@@ -25,11 +25,10 @@ fetch(`${process.env.API_URL}/graphql`, {
   .then(result => {
     // here we're filtering out any type information unrelated to unions or interfaces
     console.log('result: ', result)
-    const filteredData = result.data.__schema.types.filter(
+    result.data.__schema.types = result.data.__schema.types.filter(
       type => type.possibleTypes !== null,
     );
-    result.data.__schema.types = filteredData;
-    fs.writeFile('./src/api/fragmentTypes.json', JSON.stringify(result.data), err => {
+    fs.writeFile('./src/types/fragmentTypes.json', JSON.stringify(result.data), err => {
       if (err) {
         console.error('Error writing fragmentTypes file', err);
       } else {
